Upgrade to .NET 10 Update Nuget packages Fix broken code

+++ BookCollector/ViewModels/Groupings/AuthorsViewModel.cs
@@ -7,6 +7,7 @@ using BookCollector.ViewModels.Popups;
 using BookCollector.Views.Author;
 using BookCollector.Views.Popups;
 using CommunityToolkit.Maui.Core.Extensions;
+using CommunityToolkit.Maui.Extensions;
 using CommunityToolkit.Maui.Views;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
@@ -24,13 +25,12 @@ namespace BookCollector.ViewModels.Groupings
         public string? totalAuthorsString;
 
         private bool ShowHiddenAuthors { get; set; }
-        private bool AuthorLastNameChecked { get; set; }
         private bool TotalBooksChecked { get; set; }
         private bool TotalPriceChecked { get; set; }
 
         public AuthorsViewModel(ContentPage view)
         {
-            _view = view;
+            View = view;
             CollectionViewHeight = DeviceHeight - DoubleMenuBar;
             InfoText = $"{AppStringResources.AuthorView_InfoText}";
             ViewTitle = AppStringResources.Authors;
@@ -49,30 +49,33 @@ namespace BookCollector.ViewModels.Groupings
                     Task.Run (async () => FullAuthorList = await FilterLists.GetAllAuthorsList(ShowHiddenAuthors) ),
                 ]);
 
-                TotalAuthorsCount = FullAuthorList.Count;
-
-                FilteredAuthorList = FullAuthorList;
-
-                foreach (var author in FullAuthorList)
+                if (FullAuthorList != null)
                 {
-                    author.SetTotalBooks(ShowHiddenBook);
+                    TotalAuthorsCount = FullAuthorList.Count;
+
+                    FilteredAuthorList = FullAuthorList;
+
+                    foreach (var author in FullAuthorList)
+                    {
+                        await author.SetTotalBooks(ShowHiddenBook);
+                    }
+
+                    Task.WaitAll(
+                    [
+                        Task.Run (async () => FilteredAuthorList = await FilterLists.SortAuthorList(FilteredAuthorList,
+                                                                                                    AuthorLastNameChecked,
+                                                                                                    TotalBooksChecked,
+                                                                                                    TotalPriceChecked,
+                                                                                                    AscendingChecked,
+                                                                                                    DescendingChecked) ),
+                    ]);
+
+                    FilteredAuthorsCount = FilteredAuthorList.Count;
+
+                    TotalAuthorsString = StringManipulation.SetTotalAuthorsString(FilteredAuthorsCount, TotalAuthorsCount);
+
+                    ShowCollectionViewFooter = FilteredAuthorsCount > 0;
                 }
-
-                Task.WaitAll(
-                [
-                    Task.Run (async () => FilteredAuthorList = await FilterLists.SortAuthorList(FilteredAuthorList,
-                                                                                                AuthorLastNameChecked,
-                                                                                                TotalBooksChecked,
-                                                                                                TotalPriceChecked,
-                                                                                                AscendingChecked,
-                                                                                                DescendingChecked) ),
-                ]);
-
-                FilteredAuthorsCount = FilteredAuthorList.Count;
-
-                TotalAuthorsString = StringManipulation.SetTotalAuthorsString(FilteredAuthorsCount, TotalAuthorsCount);
-
-                ShowCollectionViewFooter = FilteredAuthorsCount > 0;
 
                 SetIsBusyFalse();
             }
@@ -83,20 +86,23 @@ namespace BookCollector.ViewModels.Groupings
         }
 
         [RelayCommand]
-        public async Task SearchOnAuthor(string? input)
+        public void SearchOnAuthor(string? input)
         {
             SetIsBusyTrue();
 
             SearchString = input;
 
-            if (!string.IsNullOrEmpty(SearchString))
-                FilteredAuthorList = FilteredAuthorList.Where(x => x.FullName.Contains(SearchString.ToLower().Trim(), StringComparison.CurrentCultureIgnoreCase)).ToObservableCollection();
-            else
-                FilteredAuthorList = FullAuthorList;
+            if (FilteredAuthorList != null)
+            {
+                if (!string.IsNullOrEmpty(SearchString))
+                    FilteredAuthorList = FilteredAuthorList.Where(x => x.FullName.Contains(SearchString.ToLower().Trim(), StringComparison.CurrentCultureIgnoreCase)).ToObservableCollection();
+                else
+                    FilteredAuthorList = FullAuthorList;
 
-            FilteredAuthorsCount = FilteredAuthorList.Count;
+                FilteredAuthorsCount = FilteredAuthorList != null ? FilteredAuthorList.Count : 0;
 
-            TotalAuthorsString = StringManipulation.SetTotalAuthorsString(FilteredAuthorsCount, TotalAuthorsCount);
+                TotalAuthorsString = StringManipulation.SetTotalAuthorsString(FilteredAuthorsCount, TotalAuthorsCount);
+            }
 
             SetIsBusyFalse();
         }
@@ -104,23 +110,27 @@ namespace BookCollector.ViewModels.Groupings
         [RelayCommand]
         public async Task PopupMenuAuthor(Guid? input)
         {
-            var selected = FilteredAuthorList.FirstOrDefault(x => x.AuthorGuid == input);
-            string? action = await PopupMenu(selected.FullName);
-
-            switch (action)
+            if (FilteredAuthorList != null)
             {
-                case "Edit":
-                    await EditAuthor(selected);
-                    break;
+                var selected = FilteredAuthorList.FirstOrDefault(x => x.AuthorGuid == input);
 
-                case "Delete":
-                    await DeleteAuthor(selected);
-                    break;
+                if (selected != null)
+                {
+                    var action = await PopupMenu(selected.FullName);
 
-                default:
-                    break;
+                    if (!string.IsNullOrEmpty(action) && action.Equals(AppStringResources.Edit))
+                    {
+                        await EditAuthor(selected);
+                    }
+
+                    if (!string.IsNullOrEmpty(action) && action.Equals(AppStringResources.Delete))
+                    {
+                        await DeleteAuthor(selected);
+                    }
+                }
             }
         }
+            
 
         [RelayCommand]
         public async Task Refresh()
@@ -135,7 +145,7 @@ namespace BookCollector.ViewModels.Groupings
         {
             SetIsBusyTrue();
 
-            AuthorEditView view = new AuthorEditView(new AuthorModel(), $"{AppStringResources.AddNewAuthor}");
+            var view = new AuthorEditView(new AuthorModel(), $"{AppStringResources.AddNewAuthor}");
 
             await Shell.Current.Navigation.PushAsync(view);
 
@@ -147,9 +157,11 @@ namespace BookCollector.ViewModels.Groupings
         {
             SetIsBusyTrue();
 
-            AuthorEditView view = new AuthorEditView(selected, selected.FullName);
-            AuthorEditViewModel bindingContext = new AuthorEditViewModel(selected, view);
-            bindingContext.ViewTitle = $"{AppStringResources.EditAuthor}";
+            var view = new AuthorEditView(selected, selected.FullName);
+            var bindingContext = new AuthorEditViewModel(selected, view)
+            {
+                ViewTitle = $"{AppStringResources.EditAuthor}"
+            };
             view.BindingContext = bindingContext;
 
             await Shell.Current.Navigation.PushAsync(view);
@@ -160,7 +172,7 @@ namespace BookCollector.ViewModels.Groupings
         [RelayCommand]
         public async Task DeleteAuthor(AuthorModel selected)
         {
-            bool answer = await DeleteCheck(selected.FullName);
+            var answer = await DeleteCheck(selected.FullName);
 
             if (answer)
             {
@@ -197,23 +209,26 @@ namespace BookCollector.ViewModels.Groupings
         [RelayCommand]
         public async Task SortPopup()
         {
-            var popup = new SortPopup();
-            SortPopupViewModel viewModel = new SortPopupViewModel(popup, ViewTitle)
+            if (!string.IsNullOrEmpty(ViewTitle))
             {
-                AuthorLastNameVisible = true,
-                AuthorLastNameChecked = AuthorLastNameChecked,
-                TotalBooksVisible = true,
-                TotalBooksChecked = TotalBooksChecked,
-                TotalPriceVisible = true,
-                TotalPriceChecked = TotalPriceChecked,
-                AscendingChecked = AscendingChecked,
-                DescendingChecked = DescendingChecked,
-            };
+                var popup = new SortPopup();
+                var viewModel = new SortPopupViewModel(popup, ViewTitle)
+                {
+                    AuthorLastNameVisible = true,
+                    AuthorLastNameChecked = AuthorLastNameChecked,
+                    TotalBooksVisible = true,
+                    TotalBooksChecked = TotalBooksChecked,
+                    TotalPriceVisible = true,
+                    TotalPriceChecked = TotalPriceChecked,
+                    AscendingChecked = AscendingChecked,
+                    DescendingChecked = DescendingChecked,
+                };
 
-            popup.BindingContext = viewModel;
+                popup.BindingContext = viewModel;
 
-            await _view.ShowPopupAsync(popup);
-            await SetViewModelData();
+                await View.ShowPopupAsync(popup);
+                await SetViewModelData();
+            }
         }
 
         private void GetPreferences()
