+++ BookCollector/ViewModels/Groupings/CollectionsViewModel.cs
@@ -8,6 +8,7 @@ using BookCollector.Views.Collection;
 using BookCollector.Views.Groupings;
 using BookCollector.Views.Popups;
 using CommunityToolkit.Maui.Core.Extensions;
+using CommunityToolkit.Maui.Extensions;
 using CommunityToolkit.Maui.Views;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
@@ -31,7 +32,7 @@ namespace BookCollector.ViewModels.Groupings
 
         public CollectionsViewModel(ContentPage view)
         {
-            _view = view;
+            View = view;
 
             CollectionViewHeight = DeviceHeight - DoubleMenuBar;
             InfoText = $"{AppStringResources.CollectionView_InfoText}";
@@ -45,37 +46,38 @@ namespace BookCollector.ViewModels.Groupings
 
                 GetPreferences();
 
-
                 Task.WaitAll(
                 [
                     Task.Run (async () => FullCollectionList = await FilterLists.GetAllCollectionsList(ShowHiddenCollections) ),
                 ]);
 
-                TotalCollectionsCount = FullCollectionList.Count;
-
-                FilteredCollectionList = FullCollectionList;
-
-                foreach (var collection in FullCollectionList)
+                if (FullCollectionList != null)
                 {
-                    collection.SetTotalBooks(ShowHiddenBook);
+                    TotalCollectionsCount = FullCollectionList.Count;
+
+                    FilteredCollectionList = FullCollectionList;
+
+                    foreach (var collection in FullCollectionList)
+                    {
+                        await collection.SetTotalBooks(ShowHiddenBook);
+                    }
+
+                    Task.WaitAll(
+                    [
+                        Task.Run (async () => FilteredCollectionList = await FilterLists.SortCollectionsList(FilteredCollectionList,
+                                                                                                             CollectionNameChecked,
+                                                                                                             TotalBooksChecked,
+                                                                                                             TotalPriceChecked,
+                                                                                                             AscendingChecked,
+                                                                                                             DescendingChecked) ),
+                    ]);
+
+                    FilteredCollectionsCount = FilteredCollectionList.Count;
+
+                    TotalCollectionsString = StringManipulation.SetTotalCollectionsString(FilteredCollectionsCount, TotalCollectionsCount);
+
+                    ShowCollectionViewFooter = FilteredCollectionsCount > 0;
                 }
-
-                Task.WaitAll(
-                [
-                    Task.Run (async () => FilteredCollectionList = await FilterLists.SortCollectionsList(FilteredCollectionList,
-                                                                                                         CollectionNameChecked,
-                                                                                                         TotalBooksChecked,
-                                                                                                         TotalPriceChecked,
-                                                                                                         AscendingChecked,
-                                                                                                         DescendingChecked) ),
-                ]);
-
-                FilteredCollectionsCount = FilteredCollectionList.Count;
-
-                TotalCollectionsString = StringManipulation.SetTotalCollectionsString(FilteredCollectionsCount, TotalCollectionsCount);
-
-                ShowCollectionViewFooter = FilteredCollectionsCount > 0;
-
 
                 SetIsBusyFalse();
             }
@@ -86,20 +88,23 @@ namespace BookCollector.ViewModels.Groupings
         }
 
         [RelayCommand]
-        public async Task SearchOnCollection(string? input)
+        public void SearchOnCollection(string? input)
         {
             SetIsBusyTrue();
 
             SearchString = input;
 
-            if (!string.IsNullOrEmpty(SearchString))
-                FilteredCollectionList = FilteredCollectionList.Where(x => x.CollectionName.Contains(SearchString.ToLower().Trim(), StringComparison.CurrentCultureIgnoreCase)).ToObservableCollection();
-            else
-                FilteredCollectionList = FullCollectionList;
+            if (FilteredCollectionList != null)
+            {
+                if (!string.IsNullOrEmpty(SearchString))
+                    FilteredCollectionList = FilteredCollectionList.Where(x => !string.IsNullOrEmpty(x.CollectionName) && x.CollectionName.Contains(SearchString.ToLower().Trim(), StringComparison.CurrentCultureIgnoreCase)).ToObservableCollection();
+                else
+                    FilteredCollectionList = FullCollectionList;
 
-            FilteredCollectionsCount = FilteredCollectionList.Count;
+                FilteredCollectionsCount = FilteredCollectionList != null ? FilteredCollectionList.Count : 0;
 
-            TotalCollectionsString = StringManipulation.SetTotalCollectionsString(FilteredCollectionsCount, TotalCollectionsCount);
+                TotalCollectionsString = StringManipulation.SetTotalCollectionsString(FilteredCollectionsCount, TotalCollectionsCount);
+            }
 
             SetIsBusyFalse();
         }
@@ -107,21 +112,24 @@ namespace BookCollector.ViewModels.Groupings
         [RelayCommand]
         public async Task PopupMenuCollection(Guid? input)
         {
-            var selected = FilteredCollectionList.FirstOrDefault(x => x.CollectionGuid == input);
-            string? action = await PopupMenu(selected.CollectionName);
-
-            switch (action)
+            if (FilteredCollectionList != null)
             {
-                case "Edit":
-                    await EditCollection(selected);
-                    break;
+                var selected = FilteredCollectionList.FirstOrDefault(x => x.CollectionGuid == input);
 
-                case "Delete":
-                    await DeleteCollection(selected);
-                    break;
+                if (selected != null && !string.IsNullOrEmpty(selected.CollectionName))
+                {
+                    var action = await PopupMenu(selected.CollectionName);
 
-                default:
-                    break;
+                    if (!string.IsNullOrEmpty(action) && action.Equals(AppStringResources.Edit))
+                    {
+                        await EditCollection(selected);
+                    }
+
+                    if (!string.IsNullOrEmpty(action) && action.Equals(AppStringResources.Delete))
+                    {
+                        await DeleteCollection(selected);
+                    }
+                }
             }
         }
 
@@ -138,7 +146,7 @@ namespace BookCollector.ViewModels.Groupings
         {
             SetIsBusyTrue();
 
-            CollectionEditView view = new CollectionEditView(new CollectionModel(), $"{AppStringResources.AddNewCollection}", true);
+            var view = new CollectionEditView(new CollectionModel(), $"{AppStringResources.AddNewCollection}", true);
 
             await Shell.Current.Navigation.PushAsync(view);
 
@@ -151,7 +159,7 @@ namespace BookCollector.ViewModels.Groupings
         {
             SetIsBusyTrue();
 
-            CollectionEditView view = new CollectionEditView(selected, $"{AppStringResources.EditCollection}", true);
+            var view = new CollectionEditView(selected, $"{AppStringResources.EditCollection}", true);
 
             await Shell.Current.Navigation.PushAsync(view);
 
@@ -161,60 +169,66 @@ namespace BookCollector.ViewModels.Groupings
         [RelayCommand]
         public async Task DeleteCollection(CollectionModel selected)
         {
-            bool answer = await DeleteCheck(selected.CollectionName);
-
-            if (answer)
+            if (!string.IsNullOrEmpty(selected.CollectionName))
             {
-                try
+                var answer = await DeleteCheck(selected.CollectionName);
+
+                if (answer)
                 {
-                    SetIsBusyTrue();
-
-                    if (TestData.UseTestData)
+                    try
                     {
-                        TestData.DeleteCollection(selected);
+                        SetIsBusyTrue();
+
+                        if (TestData.UseTestData)
+                        {
+                            TestData.DeleteCollection(selected);
+                        }
+                        else
+                        {
+
+                        }
+
+                        await ConfirmDelete(selected.CollectionName);
+
+                        await SetViewModelData();
+
+                        SetIsBusyFalse();
                     }
-                    else
+                    catch (Exception ex)
                     {
-
+                        await CanceledAction();
                     }
-
-                    await ConfirmDelete(selected.CollectionName);
-
-                    await SetViewModelData();
-
-                    SetIsBusyFalse();
                 }
-                catch (Exception ex)
+                else
                 {
                     await CanceledAction();
                 }
-            }
-            else
-            {
-                await CanceledAction();
             }
         }
 
         [RelayCommand]
         public async Task SortPopup()
         {
-            var popup = new SortPopup();
-            SortPopupViewModel viewModel = new SortPopupViewModel(popup, ViewTitle)
+            if (!string.IsNullOrEmpty(ViewTitle))
             {
-                CollectionNameVisible = true,
-                CollectionNameChecked = CollectionNameChecked,
-                TotalBooksVisible = true,
-                TotalBooksChecked = TotalBooksChecked,
-                TotalPriceVisible = true,
-                TotalPriceChecked = TotalPriceChecked,
-                AscendingChecked = AscendingChecked,
-                DescendingChecked = DescendingChecked,
-            };
+                var popup = new SortPopup();
+                var viewModel = new SortPopupViewModel(popup, ViewTitle)
+                {
+                    CollectionNameVisible = true,
+                    CollectionNameChecked = CollectionNameChecked,
+                    TotalBooksVisible = true,
+                    TotalBooksChecked = TotalBooksChecked,
+                    TotalPriceVisible = true,
+                    TotalPriceChecked = TotalPriceChecked,
+                    AscendingChecked = AscendingChecked,
+                    DescendingChecked = DescendingChecked,
+                };
 
-            popup.BindingContext = viewModel;
+                popup.BindingContext = viewModel;
 
-            await _view.ShowPopupAsync(popup);
-            await SetViewModelData();
+                await View.ShowPopupAsync(popup);
+                await SetViewModelData();
+            }
         }
 
         private void GetPreferences()
